api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
python-dotenv==1.0.0
azure-communication-email==1.0.0
requests==2.31.0
python-docx==0.8.11
PyPDF2==3.0.1
python-multipart==0.0.6
werkzeug==2.3.7

# Additional backend dependencies
azure-storage-blob==12.19.0
google-generativeai==0.3.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
redis==5.0.1
aiohttp==3.9.1
asyncio==3.4.3
